Fix for devices with /tmp mounted noexec, take 2
In the Tizen emulator image bundled with Tizen SDK 2.2, the /tmp file
system is mounted with the noexec option. This means that the
tizen-app.sh cannot be directly executed from there. This naturally
breaks all functionality that depends on tizen-app.sh.

This patch fixes the issue by prefixing device-executed scripts with an
explicit interpreter.

I didn't make the interpreter user-configurable, as I'm not sure if this
should be exposed to the user.